huella_carbono/backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow>=11.0.0
pandas==2.2.2
openpyxl==3.1.2
reportlab==4.2.0
django-filter==24.2
drf-spectacular==0.27.2
gunicorn==22.0.0
whitenoise==6.6.0
python-decouple==3.8
